pipeline/requirements.txt
anthropic>=0.40.0
google-generativeai>=0.8.0
pyyaml>=6.0
